Reorganizando la estructura de paquetes de la aplicación para crear el módulo de anuncios
