+++ src/main/webapp/WEB-INF/tiles/anuncios/definitions/tile-definition.xml
@@ -30,7 +30,7 @@
 	</definition>
 	
 	<definition name="welcome" extends="baseLayout">
-		<put-attribute name="title" value="Products" />
+		<put-attribute name="title" value="Web FDI" />
 		<put-attribute name="heading" value="Página principal" />
 		<put-attribute name="tagline" value="Seleccione una opción en el menú superior" />
 		<!-- <put-attribute name="content" value="/WEB-INF/jsp/anuncios/basico.jsp" /> -->
